+++ CHANGES.rst
@@ -4,6 +4,9 @@ CHANGES
 2.0 (unreleased)
 ----------------
 
+- Add support for Python 3
+  [pbauer]
+
 - Drop support for Zope versions older than 4.0
 
 - Fix test to work with Zope 4, where ``makeClassForTemplate`` was removed and replaced with ``SimpleViewClass``.