+++ CHANGES.rst
@@ -1,6 +1,22 @@
 CHANGES
 =======
 
+2.0.2 (unreleased)
+------------------
+
+Breaking changes:
+
+- *add item here*
+
+New features:
+
+- *add item here*
+
+Bug fixes:
+
+- *add item here*
+
+
 2.0.1 (2020-03-21)
 ------------------